requirements.txt
discord.py[voice]
mutagen
psutil
youtube_dl
aiosqlite
feedparser
python-minesweeper
python-dateutil
appdirs
